Implement 'Delete' button function by adding an eventListener (to the drag button for now, I'll change the button id later
